🐛 Fix contrib sitemap
The /sitemap.xml was broken, which is a critical component for static
site generation, both for this site.

It turns out that the `pkg.workspace` field was in the form of a
relative path specifier which causes an error when you try and
substitute it in for the `/:workspace` route.

This removes the string `"./"` from the front of any substitution.