🎉 Add @effection-contrib/timebox
Adding time constraints to any operation without having to change it at
all is one of the most powerful demonstrations of how Effection
composes. It is also needed to implement convergent assertions which
we will need in many different locations, and which we may want to
release as a library in of itself.

This takes the essence of all the timeout implementations over the
last 5 years and puts them into generally available "timebox" package.
A key decision that makes this implementation more flexible than
earlier implementations is it does not throw a "TimeoutError" when the
operation exceeds the deadline. Rather, it returns a discriminated
union indicating whether it completed successfuly, or if it timed out.